XSS/SQL Injection対策を実施、CSRFについては不明のため、一旦確認をお願いいたします #4

+++ model/cart.php
@@ -3,55 +3,22 @@ require_once MODEL_PATH . 'functions.php';
 require_once MODEL_PATH . 'db.php';
 
 function get_user_carts($db, $user_id){
-  $sql = "
-    SELECT
-      items.item_id,
-      items.name,
-      items.price,
-      items.stock,
-      items.status,
-      items.image,
-      carts.cart_id,
-      carts.user_id,
-      carts.amount
-    FROM
-      carts
-    JOIN
-      items
-    ON
-      carts.item_id = items.item_id
-    WHERE
-      carts.user_id = {$user_id}
-  ";
-  return fetch_all_query($db, $sql);
+  $sql = 'SELECT items.item_id, items.name, items.price, items.stock,
+          items.status, items.image, carts.cart_id, carts.user_id, carts.amount
+          FROM carts JOIN items ON carts.item_id = items.item_id WHERE carts.user_id=?';
+  $params = array($user_id);
+
+  return fetch_all_query($db, $sql, $params);
 }
 
 function get_user_cart($db, $user_id, $item_id){
-  $sql = "
-    SELECT
-      items.item_id,
-      items.name,
-      items.price,
-      items.stock,
-      items.status,
-      items.image,
-      carts.cart_id,
-      carts.user_id,
-      carts.amount
-    FROM
-      carts
-    JOIN
-      items
-    ON
-      carts.item_id = items.item_id
-    WHERE
-      carts.user_id = {$user_id}
-    AND
-      items.item_id = {$item_id}
-  ";
+  $sql = 'SELECT items.item_id, items.name, items.price, items.stock, 
+          items.status, items.image, carts.cart_id, carts.user_id, carts.amount
+          FROM carts JOIN items ON carts.item_id = items.item_id
+          WHERE carts.user_id=? AND items.item_id=?';
+  $params = array($user_id, $item_id);
 
-  return fetch_query($db, $sql);
-
+  return fetch_query($db, $sql, $params);
 }
 
 function add_cart($db, $user_id, $item_id ) {
@@ -63,42 +30,24 @@ function add_cart($db, $user_id, $item_id ) {
 }
 
 function insert_cart($db, $user_id, $item_id, $amount = 1){
-  $sql = "
-    INSERT INTO
-      carts(
-        item_id,
-        user_id,
-        amount
-      )
-    VALUES({$item_id}, {$user_id}, {$amount})
-  ";
+  $sql = 'INSERT INTO carts SET item_id=?, user_id=?, amount=?';
+  $params = array($item_id, $user_id, $amount);
 
-  return execute_query($db, $sql);
+  return execute_query($db, $sql, $params);
 }
 
 function update_cart_amount($db, $cart_id, $amount){
-  $sql = "
-    UPDATE
-      carts
-    SET
-      amount = {$amount}
-    WHERE
-      cart_id = {$cart_id}
-    LIMIT 1
-  ";
-  return execute_query($db, $sql);
+  $sql = 'UPDATE carts SET amount=? WHERE cart_id=? LIMIT 1';
+  $params = array($amount, $cart_id);
+
+  return execute_query($db, $sql, $params);
 }
 
 function delete_cart($db, $cart_id){
-  $sql = "
-    DELETE FROM
-      carts
-    WHERE
-      cart_id = {$cart_id}
-    LIMIT 1
-  ";
+  $sql = 'DELETE FROM carts WHERE cart_id=? LIMIT 1';
+  $params = array($cart_id);
 
-  return execute_query($db, $sql);
+  return execute_query($db, $sql, $params);
 }
 
 function purchase_carts($db, $carts){
@@ -119,12 +68,8 @@ function purchase_carts($db, $carts){
 }
 
 function delete_user_carts($db, $user_id){
-  $sql = "
-    DELETE FROM
-      carts
-    WHERE
-      user_id = {$user_id}
-  ";
+  $sql = 'DELETE FROM carts WHERE user_id=?';
+  $params = array($user_id);
 
   execute_query($db, $sql);
 }
